put SSH in a deeper folder as it won't be touched often
